+++ src/main/java/effectiveMobile/bank/util/dto/PersonListItemDto.java
@@ -5,6 +5,7 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.Date;
 
@@ -17,9 +18,9 @@ public class PersonListItemDto {
     private String email;
     private LocalDateTime birthday;
     private String phoneNumber;
-    private double amount;
+    private BigDecimal amount;
 
-    public static PersonListItemDto toDto(Person person, double amount) {
+    public static PersonListItemDto toDto(Person person, BigDecimal amount) {
         return PersonListItemDto.builder()
                 .login(person.getLogin())
                 .fullname(person.getFullname())
